+++ forge/src/main/java/net/blay09/mods/balm/forge/entity/ForgeBalmEntities.java
@@ -47,13 +47,13 @@ public class ForgeBalmEntities implements BalmEntities {
     @Override
     public <T extends LivingEntity> DeferredObject<EntityType<T>> registerEntity(ResourceLocation identifier, EntityType.Builder<T> typeBuilder, Supplier<AttributeSupplier.Builder> attributeBuilder) {
         final DeferredRegister<EntityType<?>> register = DeferredRegisters.get(ForgeRegistries.ENTITIES, identifier.getNamespace());
-        final RegistryObject<EntityType<T>> registryObject = register.register(identifier.getPath(), () -> typeBuilder.build(identifier.toString()));
         final Registrations registrations = getActiveRegistrations();
-        return new DeferredObject<>(identifier, () -> {
-            EntityType<T> entityType = registryObject.get();
+        final RegistryObject<EntityType<T>> registryObject = register.register(identifier.getPath(), () -> {
+            EntityType<T> entityType = typeBuilder.build(identifier.toString());
             registrations.attributeSuppliers.put(entityType, attributeBuilder.get().build());
             return entityType;
-        }, registryObject::isPresent);
+        });
+        return new DeferredObject<>(identifier, registryObject, registryObject::isPresent);
     }
 
     public void register() {